CPU-separation: isolate `PerThreadData`
Put common stuff in common code, and the rest inside cpu-specific file.
For cpu it's only `effective_freq_mhz`
