+++ framework/device/test_data_device.h
@@ -1,0 +1,15 @@
+/*
+ * Copyright 2025 Intel Corporation.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef INC_TEST_DATA_DEVICE_H
+#define INC_TEST_DATA_DEVICE_H
+
+#include "sandstone_config.h"
+
+#if SANDSTONE_DEVICE_CPU
+#include <device/cpu/test_data_cpu.h>
+#endif
+
+#endif /* INC_TEST_DATA_DEVICE_H */
